controllog-viz: address PR review round 5 (empty postings, split eval_review)
P2 — treat empty postings files as no postings (matsonj):
  An empty (zero-byte) postings.jsonl was still sent to read_json_auto, which
  can't infer the contract columns and failed event-only runs with "Referenced
  column posting_id not found". The reader now drops zero-byte files (events
  and postings) before building the views, falling back to the typed empty
  postings view. Added test_empty_postings_file_treated_as_no_postings.

P2 — split the eval review module (matsonj):
  eval_review.py (~1200 lines, four concerns) is now a package with a focused,
  testable boundary per concern:
    - model.py  — data model + evaluation_result extraction + text helpers
    - trace.py  — conversation-trace parsing (Chat Completions + Responses API)
    - assets.py — inline CSS/JS
    - page.py   — HTML assembly (stats bar, filters, cards, panels)
  Public API (has_eval_results, generate_eval_review) re-exported from the
  package __init__, so callers/tests are unchanged.

58 tests. Verified review/dashboard vs md:agentic_sql_logs.
