+++ projects/controllog-viz/src/controllog_viz/reader.py
@@ -107,10 +107,16 @@ def _resolve_globs(source: str) -> tuple[str, str]:
     return f"{base}/**/events.jsonl", f"{base}/**/postings.jsonl"
 
 
+def _nonempty(paths: list[str]) -> list[str]:
+    """Drop zero-byte files — read_json_auto cannot infer the contract columns from an
+    empty JSONL file, which would otherwise fail with 'Referenced column ... not found'."""
+    return [p for p in paths if Path(p).stat().st_size > 0]
+
+
 def _attach_jsonl(con: duckdb.DuckDBPyConnection, source: str) -> None:
     events_glob, postings_glob = _resolve_globs(source)
 
-    event_files = _glob.glob(events_glob, recursive=True)
+    event_files = _nonempty(_glob.glob(events_glob, recursive=True))
     if not event_files:
         raise FileNotFoundError(
             f"No controllog events found for source {source!r} "
@@ -127,7 +133,7 @@ def _attach_jsonl(con: duckdb.DuckDBPyConnection, source: str) -> None:
         "ORDER BY ingest_time DESC) = 1"
     )
 
-    posting_files = _glob.glob(postings_glob, recursive=True)
+    posting_files = _nonempty(_glob.glob(postings_glob, recursive=True))
     if posting_files:
         con.execute(
             f"CREATE TEMP VIEW postings AS SELECT {_POSTING_SELECT} "
